+++ src/main/java/com/qa/data/visualization/repositories/ClassToolsDailyActivityRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface ClassToolsDailyActivityRepository extends CrudRepository<ClassToolsDailyActivity, Integer> {
+public interface ClassToolsDailyActivityRepository extends CrudRepository<ClassToolsDailyActivity, Long> {
 
     @Query("select c from ClassToolsDailyActivity c where c.type = :type and c.time * 1000 >= 1325347200000")
     public List<ClassToolsDailyActivity> findByType(@Param("type") String type);
