+++ src/main/java/com/qa/data/visualization/repositories/LastDebugRepository.java
@@ -1,8 +1,7 @@
 package com.qa.data.visualization.repositories;
 
 import com.qa.data.visualization.entities.LastDebug;
-import com.qa.data.visualization.entities.LastError;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LastDebugRepository extends CrudRepository<LastDebug, Integer> {
+public interface LastDebugRepository extends CrudRepository<LastDebug, Long> {
 }
